Split up Snowflake tests and utils
