+++ products/batch_exports/backend/tests/temporal/destinations/snowflake/test_snowflake_batch_export_workflow.py
@@ -1,26 +1,17 @@
 import asyncio
 import dataclasses
 import datetime as dt
-import gzip
 import io
 import json
-import operator
 import os
-import re
-import tempfile
-import typing as t
 import unittest.mock
 import uuid
-from collections import deque
 from uuid import uuid4
 
 import paramiko
 import pytest
-import pytest_asyncio
 import responses
-import snowflake.connector
 from django.test import override_settings
-from requests.models import PreparedRequest
 from temporalio import activity
 from temporalio.client import WorkflowFailureError
 from temporalio.common import RetryPolicy
@@ -34,11 +25,8 @@ from posthog.batch_exports.service import (
     BatchExportModel,
     BatchExportSchema,
 )
-from posthog.temporal.common.clickhouse import ClickHouseClient
 from posthog.temporal.tests.utils.events import generate_test_events_in_clickhouse
 from posthog.temporal.tests.utils.models import (
-    acreate_batch_export,
-    adelete_batch_export,
     afetch_batch_export_runs,
 )
 from posthog.temporal.tests.utils.persons import (
@@ -57,22 +45,45 @@ from products.batch_exports.backend.temporal.destinations.snowflake_batch_export
     SnowflakeInsertInputs,
     insert_into_snowflake_activity,
     load_private_key,
-    snowflake_default_fields,
 )
-from products.batch_exports.backend.temporal.record_batch_model import SessionsRecordBatchModel
 from products.batch_exports.backend.temporal.spmc import (
-    Producer,
-    RecordBatchQueue,
     RecordBatchTaskError,
+)
+from products.batch_exports.backend.tests.temporal.destinations.snowflake.utils import (
+    FakeSnowflakeConnection,
+    add_mock_snowflake_api,
+    assert_clickhouse_records_in_snowflake,
+    contains_queries_in_order,
 )
 from products.batch_exports.backend.tests.temporal.utils import (
     FlakyClickHouseClient,
-    get_record_batch_from_queue,
     mocked_start_batch_export_run,
-    remove_duplicates_from_records,
 )
 
 pytestmark = [pytest.mark.asyncio, pytest.mark.django_db]
+
+
+TEST_TIME = dt.datetime.now(dt.UTC).replace(hour=0, minute=0, second=0, microsecond=0)
+
+REQUIRED_ENV_VARS = (
+    "SNOWFLAKE_WAREHOUSE",
+    "SNOWFLAKE_ACCOUNT",
+    "SNOWFLAKE_USERNAME",
+)
+
+
+def snowflake_env_vars_are_set():
+    if not all(env_var in os.environ for env_var in REQUIRED_ENV_VARS):
+        return False
+    if "SNOWFLAKE_PASSWORD" not in os.environ and "SNOWFLAKE_PRIVATE_KEY" not in os.environ:
+        return False
+    return True
+
+
+SKIP_IF_MISSING_REQUIRED_ENV_VARS = pytest.mark.skipif(
+    not snowflake_env_vars_are_set(),
+    reason="Snowflake required env vars are not set",
+)
 
 EXPECTED_PERSONS_BATCH_EXPORT_FIELDS = [
     "team_id",
@@ -86,351 +97,33 @@ EXPECTED_PERSONS_BATCH_EXPORT_FIELDS = [
     "is_deleted",
 ]
 
-
-TEST_TIME = dt.datetime.now(dt.UTC).replace(hour=0, minute=0, second=0, microsecond=0)
-
-
-class FakeSnowflakeCursor:
-    """A fake Snowflake cursor that can fail on PUT and COPY queries."""
-
-    def __init__(self, *args, failure_mode: str | None = None, **kwargs):
-        self._execute_calls: list[dict[str, t.Any]] = []
-        self._execute_async_calls: list[dict[str, t.Any]] = []
-        self._sfqid = 1
-        self._fail = failure_mode
-
-    @property
-    def sfqid(self):
-        current = self._sfqid
-        self._sfqid += 1
-        return current
-
-    def execute(self, query, params=None, file_stream=None):
-        self._execute_calls.append({"query": query, "params": params, "file_stream": file_stream})
-
-    def execute_async(self, query, params=None, file_stream=None):
-        self._execute_async_calls.append({"query": query, "params": params, "file_stream": file_stream})
-
-    def get_results_from_sfqid(self, query_id):
-        pass
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *args, **kwargs):
-        pass
-
-    def fetchone(self):
-        if self._fail == "put":
-            return (
-                "test",
-                "test.gz",
-                456,
-                0,
-                "NONE",
-                "GZIP",
-                "FAILED",
-                "Some error on put",
-            )
-        else:
-            return (
-                "test",
-                "test.gz",
-                456,
-                0,
-                "NONE",
-                "GZIP",
-                "UPLOADED",
-                None,
-            )
-
-    def fetchall(self):
-        if self._fail == "copy":
-            return [("test", "LOAD FAILED", 100, 99, 1, 1, "Some error on copy", 3)]
-        else:
-            return [("test", "LOADED", 100, 99, 1, 1, "Some error on copy", 3)]
-
-    def description(self):
-        return []
-
-
-class FakeSnowflakeConnection:
-    def __init__(
-        self,
-        *args,
-        failure_mode: str | None = None,
-        **kwargs,
-    ):
-        self._cursors: list[FakeSnowflakeCursor] = []
-        self._is_running = True
-        self.failure_mode = failure_mode
-
-    def cursor(self) -> FakeSnowflakeCursor:
-        cursor = FakeSnowflakeCursor(failure_mode=self.failure_mode)
-        self._cursors.append(cursor)
-        return cursor
-
-    def get_query_status_throw_if_error(self, query_id):
-        return snowflake.connector.constants.QueryStatus.SUCCESS
-
-    def is_still_running(self, status):
-        current_status = self._is_running
-        self._is_running = not current_status
-        return current_status
-
-    def close(self, *args, **kwargs):
-        pass
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *args, **kwargs):
-        pass
-
-
-def contains_queries_in_order(queries: list[str], *queries_to_find: str):
-    """Check if a list of queries contains a list of queries in order."""
-    # We use a deque to pop the queries we find off the list of queries to
-    # find, so that we can check that they are in order.
-    # Note that we use regexes to match the queries, so we can more specifically
-    # target the queries we want to find.
-    queries_to_find_deque = deque(queries_to_find)
-    for query in queries:
-        if not queries_to_find_deque:
-            # We've found all the queries we need to find.
-            return True
-        if re.match(queries_to_find_deque[0], query):
-            # We found the query we were looking for, so pop it off the deque.
-            queries_to_find_deque.popleft()
-    return not queries_to_find_deque
-
-
-def add_mock_snowflake_api(rsps: responses.RequestsMock, fail: bool | str = False):
-    # Create a crube mock of the Snowflake API that just stores the queries
-    # in a list for us to inspect.
-    #
-    # We also mock the login request, as well as the PUT file transfer
-    # request. For the latter we also store the data that was contained in
-    # the file.
-    queries = []
-    staged_files = []
-
-    def query_request_handler(request: PreparedRequest):
-        assert isinstance(request.body, bytes)
-        sql_text = json.loads(gzip.decompress(request.body))["sqlText"]
-        queries.append(sql_text)
-
-        rowset: list[tuple[t.Any, ...]] = [("test", "LOADED", 456, 192, "NONE", "GZIP", "UPLOADED", "")]
-
-        # If the query is something that looks like `PUT file:///tmp/tmp50nod7v9
-        # @%"events"` we extract the /tmp/tmp50nod7v9 and store the file
-        # contents as a string in `staged_files`.
-        if match := re.match(r"^PUT file://(?P<file_path>.*) @%(?P<table_name>.*)$", sql_text):
-            file_path = match.group("file_path")
-            with open(file_path) as f:
-                staged_files.append(f.read())
-
-            if fail == "put":
-                rowset = [
-                    (
-                        "test",
-                        "test.gz",
-                        456,
-                        0,
-                        "NONE",
-                        "GZIP",
-                        "FAILED",
-                        "Some error on put",
-                    )
-                ]
-
-        else:
-            if fail == "copy":
-                rowset = [("test", "LOAD FAILED", 100, 99, 1, 1, "Some error on copy", 3)]
-
-        return (
-            200,
-            {},
-            json.dumps(
-                {
-                    "code": None,
-                    "message": None,
-                    "success": True,
-                    "data": {
-                        "parameters": [],
-                        "rowtype": [
-                            {
-                                "name": "source",
-                                "type": "TEXT",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "target",
-                                "type": "TEXT",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "source_size",
-                                "type": "fixed",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "target_size",
-                                "type": "fixed",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "source_compression",
-                                "type": "TEXT",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "target_compression",
-                                "type": "TEXT",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "status",
-                                "type": "TEXT",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                            {
-                                "name": "message",
-                                "type": "TEXT",
-                                "length": 0,
-                                "precision": None,
-                                "scale": None,
-                                "nullable": True,
-                            },
-                        ],
-                        "rowset": rowset,
-                        "total": 1,
-                        "returned": 1,
-                        "queryId": str(uuid4()),
-                        "queryResultFormat": "json",
-                    },
-                }
-            ),
-        )
-
-    rsps.add(
-        responses.POST,
-        "https://account.snowflakecomputing.com:443/session/v1/login-request",
-        json={
-            "success": True,
-            "data": {
-                "token": "test-token",
-                "masterToken": "test-token",
-                "code": None,
-                "message": None,
-            },
+TEST_MODELS: list[BatchExportModel | BatchExportSchema | None] = [
+    BatchExportModel(
+        name="a-custom-model",
+        schema={
+            "fields": [
+                {"expression": "event", "alias": "event"},
+                {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_0)s), '')", "alias": "browser"},
+                {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_1)s), '')", "alias": "os"},
+                {"expression": "nullIf(properties, '')", "alias": "all_properties"},
+            ],
+            "values": {"hogql_val_0": "$browser", "hogql_val_1": "$os"},
         },
-    )
-    rsps.add_callback(
-        responses.POST,
-        "https://account.snowflakecomputing.com:443/queries/v1/query-request",
-        callback=query_request_handler,
-    )
-
-    return queries, staged_files
-
-
-@pytest.fixture
-def database():
-    """Generate a unique database name for tests."""
-    return f"test_batch_exports_{uuid4()}"
-
-
-@pytest.fixture
-def schema():
-    """Generate a unique schema name for tests."""
-    return f"test_batch_exports_{uuid4()}"
-
-
-@pytest.fixture
-def table_name(ateam, interval):
-    return f"test_workflow_table_{ateam.pk}_{interval}"
-
-
-@pytest.fixture
-def snowflake_config(database, schema) -> dict[str, str]:
-    """Return a Snowflake configuration dictionary to use in tests.
-
-    We set default configuration values to support tests against the Snowflake API
-    and tests that mock it.
-    """
-    warehouse = os.getenv("SNOWFLAKE_WAREHOUSE", "warehouse")
-    account = os.getenv("SNOWFLAKE_ACCOUNT", "account")
-    role = os.getenv("SNOWFLAKE_ROLE", "role")
-    username = os.getenv("SNOWFLAKE_USERNAME", "username")
-    password = os.getenv("SNOWFLAKE_PASSWORD", "password")
-    private_key = os.getenv("SNOWFLAKE_PRIVATE_KEY")
-    private_key_passphrase = os.getenv("SNOWFLAKE_PRIVATE_KEY_PASSPHRASE")
-
-    config = {
-        "user": username,
-        "warehouse": warehouse,
-        "account": account,
-        "database": database,
-        "schema": schema,
-        "role": role,
-    }
-    if private_key:
-        config["private_key"] = private_key
-        config["private_key_passphrase"] = private_key_passphrase
-        config["authentication_type"] = "keypair"
-    elif password:
-        config["password"] = password
-        config["authentication_type"] = "password"
-    else:
-        raise ValueError("Either password or private key must be set")
-    return config
-
-
-@pytest_asyncio.fixture
-async def snowflake_batch_export(ateam, table_name, snowflake_config, interval, exclude_events, temporal_client):
-    """Manage BatchExport model (and associated Temporal Schedule) for tests"""
-    destination_data = {
-        "type": "Snowflake",
-        "config": {**snowflake_config, "table_name": table_name, "exclude_events": exclude_events},
-    }
-    batch_export_data = {
-        "name": "my-production-snowflake-export",
-        "destination": destination_data,
-        "interval": interval,
-    }
-
-    batch_export = await acreate_batch_export(
-        team_id=ateam.pk,
-        name=batch_export_data["name"],
-        destination_data=batch_export_data["destination"],
-        interval=batch_export_data["interval"],
-    )
-
-    yield batch_export
-
-    await adelete_batch_export(batch_export, temporal_client)
+    ),
+    BatchExportModel(name="events", schema=None),
+    BatchExportModel(name="persons", schema=None),
+    BatchExportModel(name="sessions", schema=None),
+    {
+        "fields": [
+            {"expression": "event", "alias": "event"},
+            {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_0)s), '')", "alias": "browser"},
+            {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_1)s), '')", "alias": "os"},
+            {"expression": "nullIf(properties, '')", "alias": "all_properties"},
+        ],
+        "values": {"hogql_val_0": "$browser", "hogql_val_1": "$os"},
+    },
+    None,
+]
 
 
 class TestSnowflakeExportWorkflowMockedConnection:
@@ -885,230 +578,6 @@ class TestSnowflakeExportWorkflowErrorHandling:
         assert run.latest_error == "Cancelled"
 
 
-async def assert_clickhouse_records_in_snowflake(
-    snowflake_cursor: snowflake.connector.cursor.SnowflakeCursor,
-    clickhouse_client: ClickHouseClient,
-    table_name: str,
-    team_id: int,
-    data_interval_start: dt.datetime,
-    data_interval_end: dt.datetime,
-    exclude_events: list[str] | None = None,
-    include_events: list[str] | None = None,
-    batch_export_model: BatchExportModel | BatchExportSchema | None = None,
-    backfill_details: BackfillDetails | None = None,
-    sort_key: str = "event",
-    expected_fields: list[str] | None = None,
-    expect_duplicates: bool = False,
-    primary_key: list[str] | None = None,
-):
-    """Assert ClickHouse records are written to Snowflake table.
-
-    Arguments:
-        snowflake_cursor: A SnowflakeCursor used to read records.
-        clickhouse_client: A ClickHouseClient used to read records that are expected to be exported.
-        team_id: The ID of the team that we are testing for.
-        table_name: Snowflake table name where records are exported to.
-        data_interval_start: Start of the batch period for exported records.
-        data_interval_end: End of the batch period for exported records.
-        exclude_events: Event names to be excluded from the export.
-        include_events: Event names to be included in the export.
-        batch_export_schema: Custom schema used in the batch export.
-        expected_fields: List of fields expected to be in the destination table.
-        expect_duplicates: Whether duplicates are expected (e.g. when testing retrying logic).
-    """
-    snowflake_cursor.execute(f'SELECT * FROM "{table_name}"')
-
-    rows = snowflake_cursor.fetchall()
-
-    columns = {index: metadata.name for index, metadata in enumerate(snowflake_cursor.description)}
-    json_columns = ("properties", "person_properties", "people_set", "people_set_once", "urls")
-
-    # Rows are tuples, so we construct a dictionary using the metadata from cursor.description.
-    # We rely on the order of the columns in each row matching the order set in cursor.description.
-    # This seems to be the case, at least for now.
-    inserted_records = [
-        {
-            columns[index]: json.loads(row[index])
-            if columns[index] in json_columns and row[index] is not None
-            else row[index]
-            for index in columns.keys()
-        }
-        for row in rows
-    ]
-
-    if batch_export_model is not None:
-        if isinstance(batch_export_model, BatchExportModel):
-            model_name = batch_export_model.name
-            fields = batch_export_model.schema["fields"] if batch_export_model.schema is not None else None
-            filters = batch_export_model.filters
-            extra_query_parameters = (
-                batch_export_model.schema["values"] if batch_export_model.schema is not None else None
-            )
-        else:
-            model_name = "custom"
-            fields = batch_export_model["fields"]
-            filters = None
-            extra_query_parameters = batch_export_model["values"]
-    else:
-        model_name = "events"
-        extra_query_parameters = None
-        fields = None
-        filters = None
-
-    expected_records = []
-    queue = RecordBatchQueue()
-    if model_name == "sessions":
-        producer = Producer(model=SessionsRecordBatchModel(team_id))
-    else:
-        producer = Producer()
-
-    producer_task = await producer.start(
-        queue=queue,
-        model_name=model_name,
-        team_id=team_id,
-        full_range=(data_interval_start, data_interval_end),
-        done_ranges=[],
-        fields=fields,
-        filters=filters,
-        destination_default_fields=snowflake_default_fields(),
-        exclude_events=exclude_events,
-        include_events=include_events,
-        is_backfill=backfill_details is not None,
-        backfill_details=backfill_details,
-        extra_query_parameters=extra_query_parameters,
-    )
-    while True:
-        record_batch = await get_record_batch_from_queue(queue, producer_task)
-
-        if record_batch is None:
-            break
-
-        select = record_batch.column_names
-        if expected_fields:
-            select = expected_fields
-
-        for record in record_batch.select(select).to_pylist():
-            expected_record = {}
-
-            for k, v in record.items():
-                if k == "_inserted_at":
-                    # _inserted_at is not exported, only used for tracking progress.
-                    continue
-
-                if k in json_columns and isinstance(v, str):
-                    expected_record[k] = json.loads(v)
-                elif isinstance(v, dt.datetime):
-                    # By default, Snowflake's `TIMESTAMP` doesn't include a timezone component.
-                    expected_record[k] = v.replace(tzinfo=None)
-                elif k == "elements":
-                    # Happens transparently when uploading elements as a variant field.
-                    expected_record[k] = json.dumps(v)
-                else:
-                    expected_record[k] = v
-
-            expected_records.append(expected_record)
-
-    if expect_duplicates:
-        inserted_records = remove_duplicates_from_records(inserted_records, primary_key)
-
-    inserted_column_names = list(inserted_records[0].keys())
-    expected_column_names = list(expected_records[0].keys())
-    inserted_column_names.sort()
-    expected_column_names.sort()
-
-    # Ordering is not guaranteed, so we sort before comparing.
-    inserted_records.sort(key=operator.itemgetter(sort_key))
-    expected_records.sort(key=operator.itemgetter(sort_key))
-
-    assert inserted_column_names == expected_column_names
-    assert len(inserted_records) == len(expected_records)
-    assert inserted_records[0] == expected_records[0]
-    assert inserted_records == expected_records
-    assert len(inserted_column_names) > 0
-
-
-REQUIRED_ENV_VARS = (
-    "SNOWFLAKE_WAREHOUSE",
-    "SNOWFLAKE_ACCOUNT",
-    "SNOWFLAKE_USERNAME",
-)
-
-
-def snowflake_env_vars_are_set():
-    if not all(env_var in os.environ for env_var in REQUIRED_ENV_VARS):
-        return False
-    if "SNOWFLAKE_PASSWORD" not in os.environ and "SNOWFLAKE_PRIVATE_KEY" not in os.environ:
-        return False
-    return True
-
-
-SKIP_IF_MISSING_REQUIRED_ENV_VARS = pytest.mark.skipif(
-    not snowflake_env_vars_are_set(),
-    reason="Snowflake required env vars are not set",
-)
-
-
-@pytest.fixture
-def snowflake_cursor(snowflake_config):
-    """Manage a snowflake cursor that cleans up after we are done."""
-    password = None
-    private_key = None
-    if snowflake_config["authentication_type"] == "keypair":
-        if snowflake_config.get("private_key") is None:
-            raise ValueError("Private key is required for keypair authentication")
-
-        private_key = load_private_key(snowflake_config["private_key"], snowflake_config["private_key_passphrase"])
-    else:
-        password = snowflake_config["password"]
-
-    with snowflake.connector.connect(
-        user=snowflake_config["user"],
-        password=password,
-        role=snowflake_config["role"],
-        account=snowflake_config["account"],
-        warehouse=snowflake_config["warehouse"],
-        private_key=private_key,
-    ) as connection:
-        connection.telemetry_enabled = False
-        cursor = connection.cursor()
-        cursor.execute(f'CREATE DATABASE "{snowflake_config["database"]}"')
-        cursor.execute(f'CREATE SCHEMA "{snowflake_config["database"]}"."{snowflake_config["schema"]}"')
-        cursor.execute(f'USE SCHEMA "{snowflake_config["database"]}"."{snowflake_config["schema"]}"')
-
-        yield cursor
-
-        cursor.execute(f'DROP DATABASE IF EXISTS "{snowflake_config["database"]}" CASCADE')
-
-
-TEST_MODELS: list[BatchExportModel | BatchExportSchema | None] = [
-    BatchExportModel(
-        name="a-custom-model",
-        schema={
-            "fields": [
-                {"expression": "event", "alias": "event"},
-                {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_0)s), '')", "alias": "browser"},
-                {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_1)s), '')", "alias": "os"},
-                {"expression": "nullIf(properties, '')", "alias": "all_properties"},
-            ],
-            "values": {"hogql_val_0": "$browser", "hogql_val_1": "$os"},
-        },
-    ),
-    BatchExportModel(name="events", schema=None),
-    BatchExportModel(name="persons", schema=None),
-    BatchExportModel(name="sessions", schema=None),
-    {
-        "fields": [
-            {"expression": "event", "alias": "event"},
-            {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_0)s), '')", "alias": "browser"},
-            {"expression": "nullIf(JSONExtractString(properties, %(hogql_val_1)s), '')", "alias": "os"},
-            {"expression": "nullIf(properties, '')", "alias": "all_properties"},
-        ],
-        "values": {"hogql_val_0": "$browser", "hogql_val_1": "$os"},
-    },
-    None,
-]
-
-
 @SKIP_IF_MISSING_REQUIRED_ENV_VARS
 class TestInsertIntoSnowflakeActivity:
     @pytest.mark.parametrize("exclude_events", [None, ["test-exclude"]], indirect=True)
@@ -1352,15 +821,6 @@ class TestInsertIntoSnowflakeActivity:
         assert len(rows) == 1
         assert rows[0][0] == new_event_properties["$session_id"]
         assert rows[0][1] == dt.datetime.fromisoformat(new_event["timestamp"])
-
-    @pytest.fixture
-    def garbage_jsonl_file():
-        """Manage a JSON file with garbage data."""
-        with tempfile.NamedTemporaryFile("w+b", suffix=".jsonl", prefix="garbage_") as garbage_jsonl_file:
-            garbage_jsonl_file.write(b'{"team_id": totally not an integer}\n')
-            garbage_jsonl_file.seek(0)
-
-            yield garbage_jsonl_file.name
 
     async def test_insert_into_snowflake_activity_removes_internal_stage_files(
         self,
